Make export work with Firefox
There seems to be a race condition with Firefox 3.6. I was literally
getting Chinese characters when dragging the result into Firefox.
Moving to using Futures speeds things up enough to not bungle the
character encoding.

Also, export only text/uri-list and text/plain. This is what is needed
for Firefox and Nautilus to both recognize results as files.