Remove references to "filters" from the UI
A filter is a single piece of code that runs on a Diamond server, and does
not have a visual representation in the HyperFind interface.  The
"filters" displayed in the UI are actually higher-level constructs
consisting of a GUI for configuring a group of one or more filters.  Avoid
confusing these entities.